# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Skladi\u0161\u010de API",
    "description": "Inventory management API for tracking physical items and who holds them. All item movements are modeled as transfers between owners (people or locations).",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/owners": {
      "get": {
        "summary": "List owners",
        "tags": [
          "Owners"
        ],
        "description": "All roles. Optionally filter by type.",
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "person",
                "location"
              ]
            },
            "description": "Filter by owner type"
          }
        ],
        "responses": {
          "200": {
            "description": "List of owners",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Owner"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create owner",
        "tags": [
          "Owners"
        ],
        "description": "Manager+ only.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "type"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "type": {
                    "type": "string",
                    "enum": [
                      "person",
                      "location"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Owner created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Owner"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/items": {
      "get": {
        "summary": "List items",
        "tags": [
          "Items"
        ],
        "description": "All roles. Optionally filter by status.",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "active",
                "damaged",
                "lost",
                "removed"
              ]
            },
            "description": "Filter by item status"
          }
        ],
        "responses": {
          "200": {
            "description": "List of items",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Item"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create item",
        "tags": [
          "Items"
        ],
        "description": "Manager+ only.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Item created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Item"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Get a token from POST /api/auth/login, then pass it as: Authorization: Bearer <token>"
      }
    },
...
    "schemas": {
...
      "Owner": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "person",
              "location"
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "deleted_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
      "Item": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "image_mime": {
            "type": "string",
            "description": "MIME type of the stored image, if any"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "damaged",
              "lost",
              "removed"
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "deleted_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
...
    },
...
  }
}